Specify content type for uploaded files in S3, where one could be determined.
As far as determining the appropriate content for a given file, at first `java.nio.Files#probeContentType()` is called to see if that returns a non-null content type. If unsuccessful, `java.net.URLConnection#guessContentTypeFromName()` is used to pick a content type based on the file extension. For each file in the list of artifacts to be uploaded, the content type (if any) is recorded in a separate map. Additionally, the content type is passed in to the S3BlobStore method which calculates the presigned URL since the content-type header is part of the signed URL and needs to be known before generating the URL. Next, during the loop which uploads the artifacts to S3, the recorded content types for each file (again, if any was able to be determined for a given file) are retrieved and passed in to the RobustHTTPClient's `uploadFile` call.

In S3BlobStore, we need to use `GeneratePresignedUrlRequest` to generate the presigned URL as it allows us to specify the "Content-Type" header in the generated URL. The existing `AmazonS3ClientBuilder` does not provide this option.

For stashing of files, make sure to clear any content type that might be automatically set for a Blob, since stashing/unstashing of files is a Jenkins-internal piece of functionality and the content type of the file does not really matter there. Setting it explicitly to null ensures that the presigned URL does not include a header for whatever content type might have been set by default by the AWS SDK.

Add testcases verifying that files get the right content type set in S3 and to verify stashing and unstashing from S3.

- This requires a change to RobustHTTPClient to accept a content type parameter for an uploaded file.

Signed-off-by: Robin Verduijn <robinverduijn.github@gmail.com>